#8 Refactor directory structure.
