+++ lib/factory/entry_factory.dart
@@ -1,9 +1,9 @@
-import 'package:free_authenticator/database_entry.dart';
-import 'package:free_authenticator/entry.dart';
-import 'package:free_authenticator/entry_type.dart';
-import 'package:free_authenticator/keychain_helper.dart';
-import 'package:free_authenticator/totp.dart';
-import 'package:free_authenticator/vault.dart';
+import 'package:free_authenticator/database/database_entry.dart';
+import 'package:free_authenticator/keychain/keychain_helper.dart';
+import 'package:free_authenticator/model/entry/totp.dart';
+import 'package:free_authenticator/model/entry/vault.dart';
+import 'package:free_authenticator/model/interface/entry.dart';
+import 'package:free_authenticator/model/interface/entry_type.dart';
 
 class EntryFactory {
   static final jsonType = 'type';
