+++ lib/model/entry/entry_base.dart
