+++ lib/model/interface/entry_type.dart
