+++ lib/widget/entry/timed_password_widget.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter/material.dart';
-import 'package:free_authenticator/entry.dart';
-import 'package:free_authenticator/timer.dart';
+import 'package:free_authenticator/model/interface/entry.dart';
+import 'package:free_authenticator/widget/element/timer.dart';
 
 class TimedPasswordWidget extends StatefulWidget {
   final TimedPasswordEntry entry;
